Add an ability to load YAML files
This patch adds an ability to get content of YAML files
so it can be embedded into a web page. This is done
for convenience of the author because some code examples
could be very long. Also, I can improve it in the future in a way
that the author can use one file but reference different parts
of the file throughout an article.